feat: secure l1 block hash and blobstream address usage
